+++ service/history/execution/timer_sequence.go
@@ -82,7 +82,7 @@ type (
 
 	// TimerSequence manages user / activity timer
 	TimerSequence interface {
-		IsExpired(referenceTime time.Time, TimerSequenceID TimerSequenceID) bool
+		IsExpired(referenceTime time.Time, TimerSequenceID TimerSequenceID) (time.Duration, bool)
 
 		CreateNextUserTimer() (bool, error)
 		CreateNextActivityTimer() (bool, error)
@@ -113,11 +113,19 @@ func NewTimerSequence(
 func (t *timerSequenceImpl) IsExpired(
 	referenceTime time.Time,
 	TimerSequenceID TimerSequenceID,
-) bool {
+) (time.Duration, bool) {
 
 	// Cassandra timestamp resolution is in millisecond
 	// here we do the check in terms of second resolution.
-	return TimerSequenceID.Timestamp.Unix() <= referenceTime.Unix()
+
+	timerFireTimeInSecond := TimerSequenceID.Timestamp.Unix()
+	referenceTimeInSecond := referenceTime.Unix()
+
+	if timerFireTimeInSecond <= referenceTimeInSecond {
+		return time.Duration(referenceTimeInSecond-timerFireTimeInSecond) * time.Second, true
+	}
+
+	return 0, false
 }
 
 func (t *timerSequenceImpl) CreateNextUserTimer() (bool, error) {